+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.6.0] - 2024-04-16
+
+### Changed
+
+- 🔄 update the release process to call the command `projex git update changelog` to write the `CHANGELOG.md` file wiuth the commit list of the current branch
+- 🔄 update the tasks parameters, add `forceVtexPublish` to allow the execution with the force flag in the `vtex publish` command
+
 ## [1.4.0] - 2024-02-16
 
 ## Changed